Created rough Order class and some small edits
-Order class created (suggest to create new classes as Order is very complex now and has too many functions, as with Menu)
-Added Dhruv's Calendar and SalesReport classes (changes needed)
-Remove unnecessary code from MenuTest(likely to be added to app class)

+++ CZ2002code/src/customer/Membership.java
@@ -4,7 +4,7 @@ public class Membership {
 	private String name;
 	private int phoneNum;
 	private int membershipId;
-	public static final float discount = (float) 10.00;
+	public static final double discount = 10;
 	
 	public Membership(String name, int phoneNum, int membershipId)
 	{
